147. all working

+++ resources/item.py
@@ -1,6 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
-
+from flask_jwt_extended import jwt_required,get_jwt
 from models.item import ItemModel
 
 
@@ -38,11 +37,14 @@ class Item(Resource):
 
     @jwt_required()
     def delete(self, name):
+        claims=get_jwt()
+        if not claims['is_admin']:
+            return {"message":"Admin privilege required."},401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete()
-
-        return {"message": "item deleted"}
+            return {"message": "item deleted"}
+        return {"message": "item not found"}, 404
 
     @jwt_required()
     def put(self, name):
@@ -62,4 +64,4 @@ class Item(Resource):
 class ItemList(Resource):
 
     def get(self):
-        return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
+        return {'items': list(map(lambda x: x.json(), ItemModel.find_all()))}
